[Refactor] Clean up and split DSA indexer (#33443)
